Conflicts:
	src/main/java/com/droovy/request/UserRequestGoogleDrive.java

+++ src/main/java/com/droovy/request/UserApiRequest.java
@@ -1,15 +1,13 @@
 package com.droovy.request;
 
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.LinkedList;
 import java.util.List;
 
 import javax.ws.rs.Consumes;
-import javax.ws.rs.FormParam;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -24,6 +22,10 @@ import org.glassfish.jersey.media.multipart.FormDataParam;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sun.jersey.api.client.ClientResponse.Status;
+
+import errors.UserApplicationError;
+
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 
 @Path("request")
@@ -35,19 +37,27 @@ public class UserApiRequest {
 	UserRequest request_onedrive = new UserRequestOneDrive();
 
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/list")
-	public String getFilesList(@Context UriInfo uriInfo,@QueryParam("path") String path,@QueryParam("idUser") String idUser,@QueryParam("idFolder") String idFolder) throws JsonProcessingException {
+	public Response getFilesList(@Context UriInfo uriInfo,@QueryParam("path") String path,@QueryParam("idUser") String idUser,@QueryParam("idFolder") String idFolder,@QueryParam("getDropbox") int getDropbox,@QueryParam("getGoogleDrive") int getGoogledrive,@QueryParam("getOnedrive") int getOnedrive) throws JsonProcessingException {
+				
+		if(idUser == null || path==null || idFolder == null) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
 		
-		//TO DO : Merge les sources et fusionner si fichier identique
+		List<File> listDropbox = new LinkedList<>(), listGoogleDrive = new LinkedList<>(),listOneDrive = new LinkedList<>();
 		
-		List<File> listDropbox, listGoogleDrive,listOneDrive;
-		listDropbox = request_dropbox.getFilesList(path,idUser);
-		listGoogleDrive = request_googledrive.getFilesList(idFolder,idUser);
-		listOneDrive = request_onedrive.getFilesList(path, idUser);
+		if(getDropbox==1) {			
+			listDropbox = request_dropbox.getFilesList(path,idUser);
+		}
+		if(getGoogledrive==1) {
+			listGoogleDrive = request_googledrive.getFilesList(idFolder,idUser);
+		}
+		if(getOnedrive==1) {
+			listOneDrive = request_onedrive.getFilesList(path, idUser);
+		}
 		
 		Merger merge = new Merger();
-		
 		List<File> mergedList = merge.merge(listGoogleDrive, listDropbox, listOneDrive);
 		
 		ObjectMapper mapper = new ObjectMapper();
@@ -67,42 +77,65 @@ public class UserApiRequest {
 			output += "]";
 		}
 		
-		return output; 
+		return Response.status(Status.OK).entity(output).build();
 	}
 	
 
 	
 	@POST
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.MULTIPART_FORM_DATA)
 	@Path("/upload")
-	public String uploadFile(@FormDataParam("file") InputStream uploadInputStream, @FormDataParam("file") FormDataContentDisposition fileDetail, @QueryParam("idUser") String idUser, @QueryParam("drive") String drive) throws IOException {
+	public Response uploadFile(@FormDataParam("file") InputStream uploadInputStream, @FormDataParam("file") FormDataContentDisposition fileDetail, @FormDataParam("idUser") String idUser, @FormDataParam("drive") String drive,@FormDataParam("pathInDrive") String pathInDrive,@FormDataParam("parentId") String parentId) throws IOException {
 	
+		if(idUser == null || uploadInputStream==null || fileDetail == null || drive == null || parentId == null || pathInDrive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
 		OutputStream outputStream = new FileOutputStream(new java.io.File(fileDetail.getFileName()));
-	
+		
+		/*Sockage du fichier en local => voir si peut pas utiliser directement l'input stream*/
 		int read = 0;
-		byte[] bytes = new byte[150000000];
+		byte[] bytes = new byte[1024];
 	
 		while ((read = uploadInputStream.read(bytes)) != -1) {
 			outputStream.write(bytes, 0, read);
 		}
 
 		outputStream.close();
-		uploadInputStream.close();
+		uploadInputStream.close();		
 		
-		System.out.println("upload ");
+		File uploadedFile;
 		
-		request_dropbox.uploadFile(fileDetail.getFileName(), "/test/"+fileDetail.getFileName(), idUser);
+		if(drive.equals("dropbox")) {
+			uploadedFile=request_dropbox.uploadFile(fileDetail.getFileName(),pathInDrive, idUser,parentId);
+		}
+		else if(drive.equals("onedrive")) {
+			uploadedFile=request_onedrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser,parentId);
+		}
+		else if(drive.equals("googledrive")) {
+			uploadedFile=request_googledrive.uploadFile(fileDetail.getFileName(), pathInDrive, idUser,parentId);
+		}
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
 		
-		System.out.println("upload ");
-
-		return fileDetail.getFileName()+" "+fileDetail.getSize()+" "+fileDetail.getType()+" "+fileDetail;
+		
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(uploadedFile)+"}";
+		
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
 	}
 
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/delete")
-	public String uploadFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive) throws IOException {
+	public Response deleteFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive) throws IOException {
+		
+		if(idUser == null || path==null || idFile == null || drive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
 		
 		if(drive.equals("dropbox")) {
 			request_dropbox.removeFile(idFile, path, idUser);
@@ -113,42 +146,160 @@ public class UserApiRequest {
 		else if(drive.equals("googledrive")) {
 			request_googledrive.removeFile(idFile, path, idUser);
 		}
-		return "";
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity("{}").build();
 	}
 	
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/rename")
-	public String renameFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("name") String name) throws IOException {
+	public Response renameFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("name") String name) throws IOException {
 
+		File renamedFile;
+		
+		if(idUser == null || path==null || idFile == null || drive == null || name == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		
 		if(drive.equals("dropbox")) {
-			request_dropbox.renameFile(idFile, path, name, idUser);
+			renamedFile=request_dropbox.renameFile(idFile, path, name, idUser);
 		}
 		else if(drive.equals("onedrive")) {
-			request_onedrive.renameFile(idFile, path, name, idUser);
+			renamedFile=request_onedrive.renameFile(idFile, path, name, idUser);
 		}
 		else if(drive.equals("googledrive")) {
-			request_googledrive.renameFile(idFile, path, name, idUser);
+			renamedFile=request_googledrive.renameFile(idFile, path, name, idUser);
 		}
-		return "";
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(renamedFile)+"}";
+			
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
 	}
 	
 	@GET
-	@Produces("text/plain")
+	@Produces(MediaType.APPLICATION_JSON)
 	@Path("/move")
-	public String moveFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("idParent") String idParent, @QueryParam("pathParent") String pathParent) throws IOException {
+	public Response moveFile( @QueryParam("idUser") String idUser, @QueryParam("path") String path, @QueryParam("idFile") String idFile,@QueryParam("drive") String drive, @QueryParam("idParent") String idParent, @QueryParam("pathParent") String pathParent, @QueryParam("name") String name) throws IOException {
 
+		File movedFile;
+
+		if(idUser == null || path==null || idFile == null || drive == null || name == null || pathParent == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		
 		if(drive.equals("dropbox")) {
-			request_dropbox.moveFile(idFile, path, idParent, pathParent, idUser);
+			movedFile = request_dropbox.moveFile(idFile, path, idParent, pathParent, idUser,name);
 		}
 		else if(drive.equals("onedrive")) {
-			request_onedrive.moveFile(idFile, path, idParent, pathParent, idUser);
+			movedFile = request_onedrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
 		}
 		else if(drive.equals("googledrive")) {
-			request_googledrive.moveFile(idFile, path, idParent, pathParent, idUser);
+			movedFile = request_googledrive.moveFile(idFile, path, idParent, pathParent, idUser,name);
 		}
-		return "";
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		ObjectMapper mapper = new ObjectMapper();
+		String output = "{" + mapper.writeValueAsString(movedFile)+"}";
+			
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
+	}
+	@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/freespace")
+	public Response freeSpace( @QueryParam("idUser") String idUser, @QueryParam("drive") String drive) throws IOException {
+		
+		if(idUser == null || drive == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		String output;
+		
+		if(drive.equals("dropbox")) {
+			output=request_dropbox.freeSpaceRemaining(idUser);
+		}
+		else if(drive.equals("onedrive")) {
+			output=request_onedrive.freeSpaceRemaining(idUser);
+		}
+		else if(drive.equals("googledrive")) {
+			output=request_googledrive.freeSpaceRemaining(idUser);
+		}
+		else {
+			throw new UserApplicationError("Tell in which drive upload, example : drive=dropbox", 400);
+		}
+		
+		return Response.status(Status.OK).type(MediaType.APPLICATION_JSON).entity(output).build();
+			
 	}
 	
+	/*@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/share")
+	public Response shareFile(@QueryParam("folder") String folder,@QueryParam("permission") String permission, @QueryParam("message") String message,@QueryParam("idUser") String idUser,@QueryParam("idFile") String idFile,@QueryParam("mail") String mail){
+				
+		FilePermission permissionFile = FilePermission.READ;
+		if(permission.equals("read")) {
+			permissionFile = FilePermission.WRITE;
+		}
+		
+		
+		request_dropbox.shareFile(idUser, message, idFile, mail,permissionFile,(folder.equals("1")) ? true : false);
+		
+		return Response.status(Status.OK).entity("").build();
+		
+	}*/
+	
+	
+	@GET
+	@Produces(MediaType.APPLICATION_JSON)
+	@Path("/search")
+	public Response searchFile(@QueryParam("idUser") String idUser,@QueryParam("query") String query,@QueryParam("getDropbox") int getDropbox,@QueryParam("getGoogleDrive") int getGoogledrive,@QueryParam("getOnedrive") int getOnedrive) throws JsonProcessingException{
+	
+		if(idUser == null || query == null ) {
+			throw new UserApplicationError("At least one argument is missing", 400);
+		}
+		
+		List<File> listDropbox = new LinkedList<>(), listGoogleDrive = new LinkedList<>(),listOneDrive = new LinkedList<>();
+		
+		if(getDropbox==1) {			//Transformer en fonction ou hashmap qui regarde si token null
+			listDropbox = request_dropbox.searchFile(idUser,query);
+		}
+		if(getGoogledrive==1) {
+			listGoogleDrive = request_googledrive.searchFile(idUser,query);
+		}
+		if(getOnedrive==1) {
+			listOneDrive = request_onedrive.searchFile(idUser,query);
+		}
+		
+		listDropbox.addAll(listOneDrive);
+		listDropbox.addAll(listGoogleDrive);
+		
+		ObjectMapper mapper = new ObjectMapper();
+		
+		String output = "[";
+
+		for (File file : listDropbox) {
+			
+			output = output + mapper.writeValueAsString(file)+",";
+		}
+
+		if(listDropbox.isEmpty()) {
+			output += "]";
+		}
+		else {
+			output = output.substring(0,output.length()-1);//Retire la virgule en trop
+			output += "]";
+		}
+		
+		return Response.status(Status.OK).entity(output).build();
+	}
 	
 }
